requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langgraph
tavily-python
yfinance
pandas
plotly
pydantic
fpdf2